+++ UI/src/lib/components/header/logo.svelte
@@ -1,5 +1,5 @@
-<div>
-    <a href="/" class="font-black font-weight: 900" alt="Matthew Langley: product catalogue">
+<div class="">
+    <div href="/" class="font-black font-weight: 900">
         ML
-    </a>
+    </div>
 </div>
